move dependency to persistence
