+++ audit-server/audit_server/persistence/fastapi_dependencies.py
@@ -1,8 +1,10 @@
 """Functions to use as FastAPI dependencies for the persistence layer."""
 
+from asyncio import Lock
 from pathlib import Path
-from typing import Annotated
+from typing import Annotated, Final
 
+from anyio import Path as AsyncPath
 from fastapi import Depends
 from sqlalchemy.engine import Engine as SQLEngine
 
@@ -15,6 +17,11 @@ from server_utils.persistence.persistence_directory import PersistenceResetter
 
 _sql_engine_accessor = AppStateAccessor[SQLEngine]("sql_engine")
 _persistence_directory_root_accessor = AppStateAccessor[Path]("persistence_directory")
+
+_robot_logs_directory_init_lock = Lock()
+_robot_logs_directory_accessor = AppStateAccessor[Path]("robot_logs_directory")
+
+ROBOT_LOGS_DIRECTORY: Final = "robot_logs"
 
 
 def set_sql_engine(app_state: AppState, sql_engine: SQLEngine) -> None:
@@ -60,3 +67,20 @@ async def get_persistence_resetter(
 ) -> PersistenceResetter:
     """Get a ``PersistenceResetter`` to reset the audit-server's stored data."""
     return PersistenceResetter(directory_to_reset)
+
+
+async def get_robot_logs_directory(
+    app_state: Annotated[AppState, Depends(get_app_state)],
+    persistence_directory_root: Annotated[
+        Path, Depends(get_persistence_directory_root)
+    ],
+) -> Path:
+    """Get the directory to save the robot logs files, creating it if needed."""
+    async with _robot_logs_directory_init_lock:
+        robot_logs_dir = _robot_logs_directory_accessor.get_from(app_state)
+        if robot_logs_dir is None:
+            robot_logs_dir = persistence_directory_root / ROBOT_LOGS_DIRECTORY
+            await AsyncPath(robot_logs_dir).mkdir(exist_ok=True)
+            _robot_logs_directory_accessor.set_on(app_state, robot_logs_dir)
+
+        return robot_logs_dir
